backend/requirements.txt
Flask>=3.0.0
python-dotenv>=1.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.0
marshmallow>=3.20.0
